Create the beginnings of a "Editor Canvas Normalization" stylesheet
This commit undoes the addition of margins to P and H tags, and moves them to the editor stylesheet. That stylesheet is what gets loaded into the editing canvas already, and in such a way serves as a de-facto normalization stylesheet already.

The reason for moving paragraph margins, header margins, and header sizes there, is that we unfortunately cannot inherit those from the browser itself, if we don't style them we will inherit from wp-admin.

But by moving them to this editor stylesheet, we:

- group them all in one easy place, so they are easy to find when we _can_ refactor them out
- indicate that they are intended to override CSS bleed, and are not just opinionated styles for the sake of it
- reduce the CSS specificity of all styles involved, to be the least they can be